Added utils file; major restructuring of code
