+++ network/testing.py
@@ -1,21 +1,16 @@
-import os
 import argparse
 import numpy as np
-import matplotlib.pyplot as plt
 from datetime import datetime
 
-import torch
 from torch.utils.data import DataLoader, Subset
 import torchvision.transforms as transforms
-import torchvision.utils as utils
 
-from training import getVisualizer, getTrainingData
-from models import BaseGAN, WindowGAN, MaskedGAN, init_weights
-from models.generators import *
-from models.discriminators import *
-from datasets import PairedWindowDataset, BaseDataset, PairedFullDataset, MaskedDataset
-from visualizers import BaseGANVisualizer, PairedWindowGANVisualizer, MaskedVisualizer
-from metrics import *
+from .training import getVisualizer, getTrainingData
+from .models import BaseGAN, WindowGAN, MaskedGAN, init_weights
+from .models.generators import *
+from .models.discriminators import *
+from .datasets import PairedWindowDataset, BaseDataset, PairedFullDataset, MaskedDataset
+from utils import apply_metrics, test_metrics, Rescale
 
 
 def createModelParams(model, path, device):
